+++ backend/interfaces/api/server/handler/top_page.go
@@ -9,7 +9,7 @@ import (
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
-	"github.com/yoshihiro-shu/draft-backend/backend/internal/pkg/pager"
+	"github.com/yoshihiro-shu/draft-backend/backend/internal/pager"
 	"go.uber.org/zap"
 )
